work

+++ internal/controller/person_controller.go
@@ -2,16 +2,16 @@ package controller
 
 import (
 	"net/http"
+	"strconv"
 	"sync"
-	"uuid"
 
 	coreController "github.com/CakeForKit/rsoi-lab1/internal/common/controller"
+	"github.com/CakeForKit/rsoi-lab1/internal/common/custom_error"
 	"github.com/CakeForKit/rsoi-lab1/internal/common/model"
 	"github.com/CakeForKit/rsoi-lab1/internal/common/resolver"
 	"github.com/CakeForKit/rsoi-lab1/internal/common/utils"
 	"github.com/CakeForKit/rsoi-lab1/internal/service"
 	"github.com/gin-gonic/gin"
-	"github.com/mdobak/go-xerrors"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -58,7 +58,7 @@ func (controller *personController) RegisterHttpController(router *gin.Engine) {
 	)
 
 	personRouter.PATCH("/:personId",
-		resolver.Resolver[model.Person],
+		resolver.Resolver[model.PersonUpdate],
 		controller.update,
 	)
 
@@ -68,8 +68,11 @@ func (controller *personController) RegisterHttpController(router *gin.Engine) {
 }
 
 func (controller *personController) getById(ctx *gin.Context) {
-	id := uuid.MustParse(ctx.Param("personId"))
-	log.WithContext(ctx).Infof("UserController: GetById(id: %s): Start", id)
+	id, ok := personID(ctx)
+	if !ok {
+		return
+	}
+	log.WithContext(ctx).Infof("UserController: GetById(id: %d): Start", id)
 	if result, err := controller.service.GetById(ctx, id); err != nil {
 		log.WithError(err).WithContext(ctx).Error("UserController: GetById(): Failed")
 		utils.AbortContextWithError(ctx, err)
@@ -93,12 +96,16 @@ func (controller *personController) getAll(ctx *gin.Context) {
 func (controller *personController) create(ctx *gin.Context) {
 	log.WithContext(ctx).Info("UserController: Create(): Start")
 	user := utils.MustGetRequestBody[model.Person](ctx)
+	if user.Name == "" {
+		utils.AbortContextWithError(ctx, custom_error.IllegalArgumentError("name is required"))
+		return
+	}
 	if result, err := controller.service.Create(ctx, user); err != nil {
 		log.WithError(err).WithContext(ctx).Error("UserController: Create(): Failed")
 		utils.AbortContextWithError(ctx, err)
 	} else {
-		ctx.Header("Location", "/api/v1/persons/"+result.ID.String())
-		ctx.AbortWithStatusJSON(http.StatusCreated, result)
+		ctx.Header("Location", "/api/v1/persons/"+strconv.FormatUint(uint64(result.ID), 10))
+		ctx.Status(http.StatusCreated)
 	}
 	log.WithContext(ctx).Info("UserController: Create(): End")
 }
@@ -106,17 +113,18 @@ func (controller *personController) create(ctx *gin.Context) {
 func (controller *personController) update(ctx *gin.Context) {
 	log.WithContext(ctx).Info("UserController: Update(): Start")
 
-	id := uuid.MustParse(ctx.Param("personId"))
-	user := utils.MustGetRequestBody[model.Person](ctx)
+	id, ok := personID(ctx)
+	if !ok {
+		return
+	}
+	user := utils.MustGetRequestBody[model.PersonUpdate](ctx)
 
-	if id.String() != user.ID.String() {
-		err := xerrors.New("pathId != dto.id")
-		log.WithError(err).WithContext(ctx).Error("UserController: Update(): Failed")
-		utils.AbortContextWithError(ctx, err)
+	if user.Name == nil && user.Age == nil && user.Address == nil && user.Work == nil {
+		utils.AbortContextWithError(ctx, custom_error.IllegalArgumentError("request body is empty"))
 		return
 	}
 
-	if result, err := controller.service.Update(ctx, user); err != nil {
+	if result, err := controller.service.Update(ctx, id, user); err != nil {
 		log.WithError(err).WithContext(ctx).Error("UserController: Update(): Failed")
 		utils.AbortContextWithError(ctx, err)
 	} else {
@@ -126,13 +134,25 @@ func (controller *personController) update(ctx *gin.Context) {
 }
 
 func (controller *personController) deleteById(ctx *gin.Context) {
-	id := uuid.MustParse(ctx.Param("personId"))
-	log.WithContext(ctx).Infof("UserController: DeleteById(id: %s): Start", id)
+	id, ok := personID(ctx)
+	if !ok {
+		return
+	}
+	log.WithContext(ctx).Infof("UserController: DeleteById(id: %d): Start", id)
 	if err := controller.service.DeleteById(ctx, id); err != nil {
 		log.WithError(err).WithContext(ctx).Error("UserController: DeleteById(): Failed")
 		utils.AbortContextWithError(ctx, err)
 	} else {
-		utils.AbortContextWithOK(ctx)
+		ctx.Status(http.StatusNoContent)
 	}
 	log.WithContext(ctx).Info("UserController: DeleteById(): End")
+}
+
+func personID(ctx *gin.Context) (uint, bool) {
+	id, err := strconv.ParseUint(ctx.Param("personId"), 10, 0)
+	if err != nil || id == 0 {
+		utils.AbortContextWithError(ctx, custom_error.IllegalArgumentError("personId must be a positive integer"))
+		return 0, false
+	}
+	return uint(id), true
 }
